Announcements listview
Feature: Announcements now get fetched in real time into the home screen announcements section and the UI gets refreshed whenever a document is edited/removed or added into the announcements collection. Core functionality is explained in detail in https://pub.dev/documentation/flutterfire_ui/latest/firestore/FirestoreListView-class.html and further examples in https://codewithandrea.com/articles/firestore-pagination-list-view/ .

Announcement file in the domain layer of Home contains the class definition for the announcement class, it's parameters and factory methods for fromJson and toJson.

Announcement repository in the data layer of Home contains the logic for interacting with Firestore and it makes use of the announcement class model to create a snapshot of announcement objects. To change in which order the documents are returned in you can do it here.

In the presentation layer of Home, a new section called _AnnouncementsPreview contains all of the content for announcements and utilizes the FireStoreListView to build a listview that shows the announcements.  It currently has no limitation to how many documents it shows but that should be edited so that it only shows a handful.

Added a timestamp_formatter helper function into the utilities folder to help format Firestore Timestamp objects into a nice, readable form in the UI.

Added the relevant new fields and collections into the firestore_constants file